[30] fix: remove misc utils (#13)
